markets pdp utils
